🧬 Migrate the logic about processing response value into another single object.
